fix: Renamed click-service folder according to styleguide
