+++ src/lib/src/core/viewer-service/viewer.service.spec.ts
@@ -1,7 +1,7 @@
 import { inject, TestBed } from '@angular/core/testing';
 
 import { ViewerService } from './viewer.service';
-import { ClickService } from '../click/click.service';
+import { ClickService } from '../click-service/click.service';
 import { PageService } from './../page-service/page-service';
 import { ModeService } from './../mode-service/mode.service';
