+++ src/lib/src/viewer/osd-toolbar/osd-toolbar.component.spec.ts
@@ -1,6 +1,6 @@
 import { PageService } from '../../core/page-service/page-service';
 import { ModeService } from '../../core/mode-service/mode.service';
-import { ClickService } from '../../core/click/click.service';
+import { ClickService } from '../../core/click-service/click.service';
 import { ViewerService } from '../../core/viewer-service/viewer.service';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
